% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{apply_semisupervised_learning}
\alias{apply_semisupervised_learning}
\title{Apply Semi-supervised Learning}
\usage{
apply_semisupervised_learning(
  initial_model,
  training_data,
  all_pairs,
  features_ms1,
  features_ms2,
  confidence_threshold,
  max_iterations,
  seed
)
}
\arguments{
\item{initial_model}{Initial trained model}

\item{training_data}{Initial training data}

\item{all_pairs}{All potential pairs}

\item{features_ms1}{Features from first dataset}

\item{features_ms2}{Features from second dataset}

\item{confidence_threshold}{Confidence threshold for adding predictions to training}

\item{max_iterations}{Maximum iterations for semi-supervised learning}

\item{seed}{Random seed for reproducibility}
}
\value{
List with updated model and training data
}
\description{
Apply Semi-supervised Learning
}
