% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_combine.R
\name{auto_combine}
\alias{auto_combine}
\title{Auto Combine}
\usage{
auto_combine(
  ms1,
  ms2,
  rt_lower = -0.5,
  rt_upper = 0.5,
  mz_lower = -15,
  mz_upper = 15,
  minimum_intensity = 1000,
  iso_method = "manual",
  eps = 0.1,
  rt_iso_threshold = 0.5,
  mz_iso_threshold = 5,
  threshold = "manual",
  match_method = "unsupervised",
  smooth_method = "loess",
  multipliers = c(6, 6, 6),
  weights = c(1, 1, 1),
  keep_features = c(F, F)
)
}
\arguments{
\item{ms1}{A \code{massSight} object representing the results of a preprocessed
LC-MS experiment.}

\item{ms2}{A \code{massSight} object representing the results of a second
preprocessed LC-MS experiment.}

\item{rt_lower}{A numeric indicating the lower bound of the RT
range to be considered for aligning two metabolites.}

\item{rt_upper}{A numeric indicating the upper bound of the RT
range to be considered for aligning two metabolites.}

\item{mz_lower}{A numeric indicating the lower bound of the m/z
range to be considered for aligning two metabolites.}

\item{mz_upper}{A numeric indicating the upper bound of the m/z
range to be considered for aligning two metabolites.}

\item{minimum_intensity}{A numeric indicating the minimum intensity
to be considered for alignment.}

\item{rt_iso_threshold}{A numeric indicating the simplification
parameter for RT.}

\item{mz_iso_threshold}{A numeric indicating the simplification
parameter for m/z.}

\item{match_method}{A character indicating the initial matching method to
be used to detect inter-batch variability. Options are "unsupervised" and
"supervised".}

\item{smooth_method}{A character indicating the smoothing method to
be used. Options are "lowess", "spline", and "gaussian".}

\item{multipliers}{A numeric vector indicating the multipliers to be
used for the alignment.}

\item{weights}{A numeric vector indicating the weights to be used for
the alignment.}

\item{keep_features}{A logical vector indicating whether or not to
keep features that are not matched.}
}
\value{
A data frame containing the combined data.
}
\description{
Combines two \code{massSight} objects, resulting in a single
\code{MergedMSObject}.
}
