% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{calculate_local_mean}
\alias{calculate_local_mean}
\title{Calculate local mean for a variable within a sliding window}
\usage{
calculate_local_mean(x, y, window)
}
\arguments{
\item{x}{Vector of positions}

\item{y}{Vector of values}

\item{window}{Window size}
}
\value{
Vector of local means
}
\description{
Calculate local mean for a variable within a sliding window
}
