% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{calculate_local_sd}
\alias{calculate_local_sd}
\title{Calculate local standard deviation for a variable within a sliding window}
\usage{
calculate_local_sd(x, y, window)
}
\arguments{
\item{x}{Vector of positions}

\item{y}{Vector of values}

\item{window}{Window size}
}
\value{
Vector of local standard deviations
}
\description{
Calculate local standard deviation for a variable within a sliding window
}
