% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd2csv.R
\name{cd2csv}
\alias{cd2csv}
\title{CD-To-Csv}
\usage{
cd2csv(path, gen_id = TRUE, output_file = NULL)
}
\arguments{
\item{path}{The path to the Compound Discoverer Excel (.xlsx) file}

\item{gen_id}{If \code{TRUE}, this function will generate a unique id
for each compound based on retention time and mass to charge ratio.}

\item{output_file}{The name of the file to save the formatted data. If
\code{NULL} (default), a csv file will not be written.}
}
\value{
A dataframe that can then be used by \code{massSight}'s \code{auto_combine()}
or \code{auto_scale()}
}
\description{
Converts output from Compound Discoverer into a csv format suitable
for \code{massSight}
}
