% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{compute_match_metrics}
\alias{compute_match_metrics}
\title{Compute Match Metrics}
\usage{
compute_match_metrics(predictions, true_matches)
}
\arguments{
\item{predictions}{Predicted matches}

\item{true_matches}{Known true matches}
}
\value{
List of quality metrics
}
\description{
Compute Match Metrics
}
