% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate}
\usage{
consolidate(
  massSight_obj,
  use_rt = T,
  use_mz = T,
  rt_threshold = NULL,
  mz_threshold = NULL
)
}
\arguments{
\item{massSight_obj}{a \code{massSight} object}

\item{use_rt}{boolean if retention time should be used in identifying
metabolites that should be consolidated}

\item{use_mz}{boolean if mass to charge ration should be used in identifying
metabolites that should be consolidated}

\item{rt_threshold}{user defined retention time threshold for defining
adducts/isomers}

\item{mz_threshold}{user defined mass to charge ratio threshold for defining
adducts/isomers}
}
\value{
a \code{massSight} object with consolidated metabolites
}
\description{
Consolidate adducts/isomers with similar features into one
observation
}
