% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_isotopes.R
\name{consolidate_isotopes}
\alias{consolidate_isotopes}
\title{Group isotopes with their parent compounds and prepare for matching}
\usage{
consolidate_isotopes(ms_obj)
}
\arguments{
\item{ms_obj}{MSObject with isotope labels}
}
\value{
MSObject with consolidated isotope groups
}
\description{
Group isotopes with their parent compounds and prepare for matching
}
