% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{create_enhanced_training_data}
\alias{create_enhanced_training_data}
\title{Create Enhanced Training Data}
\usage{
create_enhanced_training_data(
  labeled_data,
  features_ms1,
  features_ms2,
  mz_thresh,
  rt_thresh
)
}
\arguments{
\item{labeled_data}{List of labeled metabolite data}

\item{features_ms1}{Features from first dataset}

\item{features_ms2}{Features from second dataset}

\item{mz_thresh}{MZ threshold for negative examples}

\item{rt_thresh}{RT threshold for negative examples}
}
\value{
Enhanced training dataset
}
\description{
Create Enhanced Training Data
}
