% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create_ms_obj}
\alias{create_ms_obj}
\title{Create MS Object}
\usage{
create_ms_obj(
  df,
  name,
  id_name = "Compound_ID",
  rt_name = "RT",
  mz_name = "MZ",
  int_name = "Intensity",
  metab_name = "Metabolite"
)
}
\arguments{
\item{df}{A data frame containing the raw data.}

\item{name}{A character indicating the name of the experiment.}

\item{id_name}{A character indicating the name of the column containing the
compound IDs.}

\item{rt_name}{A character indicating the name of the column containing the
retention times.}

\item{mz_name}{A character indicating the name of the column containing the
m/z values.}

\item{int_name}{A character indicating the name of the column containing the
intensities}

\item{metab_name}{An optional character indicating the name of the column
containing the metabolite annotations}
}
\value{
An MSObject.
}
\description{
Create an MSObject from a data frame.
}
