% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_isotopes.R
\name{detect_isotopes}
\alias{detect_isotopes}
\title{Detect and label isotopes in an MSObject}
\usage{
detect_isotopes(ms_obj, ppm_tolerance = 5, rt_tolerance = 0.1)
}
\arguments{
\item{ms_obj}{An MSObject containing mass spectrometry data}

\item{ppm_tolerance}{Mass tolerance in PPM for matching isotopes}

\item{rt_tolerance}{RT window to look for isotopes}
}
\value{
An MSObject with isotopic labels added
}
\description{
Detect and label isotopes in an MSObject
}
