% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plot.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Distribution Plot}
\usage{
distribution_plot(ms_obj, subset = "all")
}
\arguments{
\item{ms_obj}{Either a \code{MSObject}, or \code{MergedMSObject}}

\item{subset}{Whether to plot all metabolites or only isolated metabolites.
Can either be "all" or "isolated"}
}
\value{
a scatter plot of the retention times vs mass to charge ratios with
marginal histograms
}
\description{
Plot distributions of retention times and mass to charge ratios
of individual MS experiments
}
\examples{
aligned <- auto_combine(ms1, ms2)
distribution_plot(aligned, subset = "all")
}
