% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{extract_advanced_features}
\alias{extract_advanced_features}
\title{Extract Advanced Features from MS Object}
\usage{
extract_advanced_features(ms_obj)
}
\arguments{
\item{ms_obj}{A massSight object}
}
\value{
A data frame with compound IDs and extracted features
}
\description{
Extract Advanced Features from MS Object
}
