% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_intensities.R
\name{filter_intensities}
\alias{filter_intensities}
\title{Filter Intensities}
\usage{
filter_intensities(data, prevalence = 0.5)
}
\arguments{
\item{data}{Dataframe containing results from an LC-MS experiment.}

\item{prevalence}{Percent of samples required to keep a metabolite.}
}
\value{
A logical vector indicating which rows of the original dataset
should be retained
}
\description{
Filter Intensities
}
