% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{generate_potential_pairs}
\alias{generate_potential_pairs}
\title{Generate Potential Pairs}
\usage{
generate_potential_pairs(ms1, ms2, mz_thresh, rt_thresh)
}
\arguments{
\item{ms1}{First MSObject}

\item{ms2}{Second MSObject}

\item{mz_thresh}{MZ threshold in ppm}

\item{rt_thresh}{RT threshold in minutes}
}
\value{
Data frame of potential pairs within thresholds
}
\description{
Generate Potential Pairs
}
