% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_combine.R
\name{get_known_pairs}
\alias{get_known_pairs}
\title{Get pairs of known metabolites between datasets}
\usage{
get_known_pairs(merged_ms)
}
\arguments{
\item{ms1}{First MSObject}

\item{ms2}{Second MSObject}
}
\value{
Data frame of matching compound IDs for known metabolites
}
\description{
Get pairs of known metabolites between datasets
}
