% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{get_shared_metabolites}
\alias{get_shared_metabolites}
\title{Get Shared Metabolites Between Datasets}
\usage{
get_shared_metabolites(ms1, ms2)
}
\arguments{
\item{ms1}{First MSObject}

\item{ms2}{Second MSObject}
}
\value{
List containing matched metabolites and related data
}
\description{
Get Shared Metabolites Between Datasets
}
