% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msobject-class.R
\docType{class}
\name{MergedMSObject-class}
\alias{MergedMSObject-class}
\title{Class to represent merged mass spectroscopy data.}
\description{
Class to represent merged mass spectroscopy data.
}
\section{Slots}{

\describe{
\item{\code{ms1}}{A character indicating the name of the experiment.}

\item{\code{ms2}}{A data frame containing the raw data.}

\item{\code{all_matched}}{A data frame containing the scaled data.}

\item{\code{iso_matched}}{A data frame containing the scaled data.}

\item{\code{metadata}}{A logical indicating whether or not the data has
been consolidated.}

\item{\code{pre_iso_matched}}{A data frame containing the matched pairs before
initial isolation}

\item{\code{scaled_values}}{A vector of scaled values}

\item{\code{adjusted_df}}{A data frame containing the drift corrected metabolites}

\item{\code{cutoffs}}{A vector of cutoffs}

\item{\code{aligned}}{A dataframe containing the final metabolite mathces}

\item{\code{smooth_method}}{A string indicating the smooth_methoding method used.}
}}

\concept{objects}
