% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSObject.R
\docType{class}
\name{MSObject-class}
\alias{MSObject-class}
\title{Class to represent mass spectroscopy data.}
\description{
Class to represent mass spectroscopy data.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character indicating the name of the experiment.}

\item{\code{raw_df}}{A data frame containing the raw data.}

\item{\code{isolated}}{A data frame containing the isolated data.}

\item{\code{scaled_df}}{A data frame containing the scaled data.}

\item{\code{consolidated}}{A logical indicating whether or not the data has
been consolidated.}

\item{\code{metadata}}{A data frame containing the metadata.}
}}

\concept{objects}
