% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{plot_decision_boundary}
\alias{plot_decision_boundary}
\title{Plot Decision Boundary}
\usage{
plot_decision_boundary(ml_match_result)
}
\arguments{
\item{ml_match_result}{Result from ml_match function}
}
\value{
ggplot object showing decision boundary
}
\description{
Plot Decision Boundary
}
