% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot Feature Importance}
\usage{
plot_feature_importance(ml_match_result, top_n = 20)
}
\arguments{
\item{ml_match_result}{Result from ml_match function}

\item{top_n}{Number of top features to show}
}
\value{
ggplot object showing feature importance
}
\description{
Plot Feature Importance
}
