% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{predict_with_constraints}
\alias{predict_with_constraints}
\title{Predict With Constraints}
\usage{
predict_with_constraints(model, all_pairs, prob_thresh = 0.5)
}
\arguments{
\item{model}{Trained model}

\item{all_pairs}{All potential pairs}

\item{prob_thresh}{Probability threshold for matches}
}
\value{
Data frame of predicted matches with constraints applied
}
\description{
Predict With Constraints
}
