% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_match.R
\name{prepare_prediction_data}
\alias{prepare_prediction_data}
\title{Prepare Prediction Data}
\usage{
prepare_prediction_data(all_pairs, model_features)
}
\arguments{
\item{all_pairs}{All potential pairs}

\item{model_features}{Feature names from model}
}
\value{
Data frame ready for prediction
}
\description{
Prepare Prediction Data
}
