% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_combine.R
\name{evaluate_matches}
\alias{evaluate_matches}
\title{Evaluate matching quality using known metabolite pairs}
\usage{
evaluate_matches(result)
}
\arguments{
\item{result}{Result from mass_combine}
}
\value{
Numeric score between 0 and 1
}
\description{
Evaluate matching quality using known metabolite pairs
}
