% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R
\name{final_plots}
\alias{final_plots}
\title{Final Plots}
\usage{
final_plots(merged_ms_obj, rt_lim = c(-0.5, 0.5), mz_lim = c(-15, 15))
}
\arguments{
\item{merged_ms_obj}{A Merged MS Object, typically the result of \code{auto_align}}

\item{rt_lim}{RT bounds}

\item{mz_lim}{MZ bounds}
}
\value{
cowplot grid of plots
}
\description{
Final Plots
}
