% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_combine.R
\name{get_unique_matches}
\alias{get_unique_matches}
\title{Get unique 1-1 matches from mass_combine output}
\usage{
get_unique_matches(ms_object)
}
\arguments{
\item{all_matched}{Data frame containing all potential matches from mass_combine}
}
\value{
Data frame containing only unique 1-1 matches, where each feature appears only once
}
\description{
Get unique 1-1 matches from mass_combine output
}
