% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massSight-package.R
\docType{package}
\name{massSight-package}
\alias{massSight}
\alias{massSight-package}
\title{massSight: Batch Alignment and Scaling for LC-MS Data}
\description{
an `R` package for the alignment, scaling, and downstream analysis of LC-MS metabolomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/omicsEye/massSight}
  \item \url{https://omicseye.github.io/massSight/}
  \item Report bugs at \url{https://github.com/omicsEye/massSight/issues}
}

}
\author{
\strong{Maintainer}: Chiraag Gohel \email{chiraaggohel@gwu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Ali Rahnavard \email{rahnavard@gwu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
