% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{MergedMSObject-class}
\alias{MergedMSObject-class}
\title{Class to represent merged mass spectroscopy data.}
\description{
Class to represent merged mass spectroscopy data.
}
\section{Slots}{

\describe{
\item{\code{ms1}}{A character indicating the name of the experiment.}

\item{\code{ms2}}{A data frame containing the raw data.}

\item{\code{all_matched}}{A data frame containing the scaled data.}

\item{\code{iso_matched}}{A data frame containing the scaled data.}

\item{\code{metadata}}{A logical indicating whether or not the data has
been consolidated.}

\item{\code{smooth}}{A string indicating the smoothing method used.}
}}

\concept{objects}
